+++ scripts/lichessTools-init.js
@@ -146,6 +146,7 @@ lichessTools.loadTool(LiChessTools.Tools.LobbyRatingDiff);
 lichessTools.loadTool(LiChessTools.Tools.VoiceInput);
 lichessTools.loadTool(LiChessTools.Tools.BotsMenuItem);
 lichessTools.loadTool(LiChessTools.Tools.TeamStats);
+lichessTools.loadTool(LiChessTools.Tools.CalculationTrainer);
 
 
 lichessTools.loadTool(LiChessTools.Tools.HideBoardCommand);